Added MemParser class
Added MemParser class, the class to parse the initial memory data input
file. Would store the memory data in a map data structure and can
be accessed via accessor method. The key for the map is the string hex
representation of result, whereas the data value is stored in binary
string representation.